feat: productPage to TS
