+++ src/controllers/query.ts
@@ -2,10 +2,10 @@ export function getQuery() {
   return new URLSearchParams(window.location.search);
 }
 
-export function updateRout(path) {
+export function updateRout(path: string) {
   window.history.pushState({}, path, window.location.origin + path);
 }
 
-export function updateQuery(options) {
-  const params = new URLSearchParams(window.location.search);
-}
+// export function updateQuery(options) {
+//   const params = new URLSearchParams(window.location.search);
+// }
