+++ src/types.ts
@@ -58,3 +58,5 @@ export interface CheckBoxData extends Array<CheckBox> {}
 export type HTMLElementEvent<T extends HTMLElement> = Event & {
   target: T;
 };
+
+export type Input = string | number;
